removed message formatter and removed dependencies in integerations
